refactor: agrupando as pastas relacionadas ao user
